doc-gen-agent/requirements.txt
google-genai>=0.2.2
langchain>=0.3.1
langgraph>=0.2.21
python-docx>=1.1.0
pandas>=2.2.0
pypdf>=5.0.0
python-dotenv>=1.0.1
streamlit>=1.39.0
